Adding Content
Adding Content
Now that you have a feel for what you’re working with, it’s time to start adding new code to the site.

Create a banner component at the top of the page which displays the site’s title.
Create an exhibit component that displays a heading, as well as any child components. This component will wrap the components we create in the following tasks—think of it as a pretty container to logically separate different data points. It can take any form you like, from a card to a panel with multiple tabs. Make it your own!
If you get stuck, read through the intro to React tutorial found here, then take a look at the way composition is carried out in React.

When you are finished, commit and push your changes, then submit a link to your github repo below. This task should take you approximately one hour to complete.

We’ll show you an example on the next step - but make sure to give it a go yourself first!